requirements.txt
aiohttp>=3.8.0
websockets>=10.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
ccxt 